fix: avoid overlap with notch
