+++ components/AppWrapper.tsx
@@ -12,13 +12,9 @@ import {
   NavigationContainer,
 } from '@react-navigation/native';
 
-import {View, TouchableOpacity} from 'react-native';
-import FontAwesome5 from 'react-native-vector-icons/FontAwesome5';
-
 import {Launchpad, Settings, About} from '.';
 import {useStoreState} from '../store/hooks';
-import {AngInput} from './Ang/AngInput';
-import {elementStyles} from '../styles';
+import {Header} from './AppHeader';
 
 const Drawer = createDrawerNavigator();
 
@@ -26,29 +22,11 @@ const AppWrapper = (): JSX.Element => {
   const getSettings = (props: DrawerContentComponentProps) => {
     return <Settings {...props} />;
   };
+  const getHeader = ({navigation}: any) => {
+    return <Header navigation={navigation} />;
+  };
   const {darkTheme, leftHandedMode} = useStoreState(state => state);
   const currentTheme = darkTheme === true ? DarkTheme : DefaultTheme;
-
-  const menuIcon = (navigation: any) => (
-    <TouchableOpacity
-      style={{margin: 16}}
-      onPress={() => navigation.toggleDrawer()}>
-      <FontAwesome5
-        name="bars"
-        size={22}
-        color={darkTheme ? '#FFFFFF' : '#333333'}
-      />
-    </TouchableOpacity>
-  );
-
-  const bakeHeader = (navigation: any) => {
-    return (
-      <View style={elementStyles(currentTheme.colors).navigationHeader}>
-        {leftHandedMode ? <AngInput /> : menuIcon(navigation)}
-        {leftHandedMode ? menuIcon(navigation) : <AngInput />}
-      </View>
-    );
-  };
 
   return (
     <NavigationContainer theme={currentTheme}>
@@ -58,7 +36,7 @@ const AppWrapper = (): JSX.Element => {
           component={Launchpad}
           options={({navigation}) => ({
             drawerPosition: leftHandedMode === true ? 'right' : 'left',
-            header: () => bakeHeader(navigation),
+            header: () => getHeader({navigation}),
           })}
         />
         <Drawer.Screen name="About" component={About} />
